Tread lightly, she is near

+++ 0x15-file_io/0-read_textfile.c
@@ -29,7 +29,7 @@ size_t read_textfile(const char *filename, size_t letters)
 
 	close(fd);
 	free(data);
-	if (num == -1)
+	if (num == (size_t)-1)
 		return (0);
 
 	return (num);
